HIVE-29112: addressing review comments;
