+++ standalone-metastore/metastore-server/src/test/java/org/apache/hadoop/hive/metastore/properties/HMSServletTest.java
@@ -17,7 +17,6 @@
 package org.apache.hadoop.hive.metastore.properties;
 
 import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
 import java.io.BufferedReader;
 import java.io.DataOutputStream;
 import java.io.IOException;
@@ -26,7 +25,6 @@ import java.io.Reader;
 import java.net.HttpURLConnection;
 import java.net.URI;
 import java.net.URL;
-import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import java.util.Collections;
 import java.util.List;
@@ -40,16 +38,6 @@ import org.apache.hadoop.hive.metastore.PropertyServlet;
 import org.apache.hadoop.hive.metastore.annotation.MetastoreUnitTest;
 import org.apache.hadoop.hive.metastore.conf.MetastoreConf;
 import org.apache.hadoop.hive.metastore.utils.MetaStoreUtils;
-import org.apache.http.HttpEntity;
-import org.apache.http.HttpResponse;
-import org.apache.http.NameValuePair;
-import org.apache.http.client.HttpClient;
-import org.apache.http.client.methods.HttpGet;
-import org.apache.http.client.utils.URIBuilder;
-import org.apache.http.conn.ssl.NoopHostnameVerifier;
-import org.apache.http.entity.ContentType;
-import org.apache.http.impl.client.HttpClients;
-import org.apache.http.message.BasicNameValuePair;
 import org.eclipse.jetty.server.Server;
 import org.junit.Assert;
 import org.junit.Before;
@@ -97,31 +85,27 @@ public class HMSServletTest extends HMSTestBase {
     }
   }
 
-  protected static HttpClient createHttpClient(Configuration conf) {
-    SSLContext sslCtxt = clientSSLContextFactory(conf);
-    return sslCtxt != null
-      ? HttpClients.custom().setSSLContext(sslCtxt).setSSLHostnameVerifier(NoopHostnameVerifier.INSTANCE).build()
-      : HttpClients.createDefault();
+  protected static String getScheme(Configuration conf) {
+    return MetastoreConf.getBoolVar(conf, MetastoreConf.ConfVars.HTTPSERVER_USE_HTTPS) ? "https" : "http";
   }
 
   @Override
-  protected PropertyClient createClient(Configuration conf, int sport) throws Exception {
-    String path = MetastoreConf.getVar(conf, MetastoreConf.ConfVars.PROPERTIES_SERVLET_PATH);
-    String scheme = MetastoreConf.getBoolVar(conf, MetastoreConf.ConfVars.HTTPSERVER_USE_HTTPS)
-            ? "https" : "http";
-    URI uri = new URI(scheme + "://hive@localhost:" + sport + "/" + path + "/" + NS);
+  protected PropertyClient createClient(Configuration conf, int port) throws Exception {
+    String servletPath = MetastoreConf.getVar(conf, MetastoreConf.ConfVars.PROPERTIES_SERVLET_PATH);
+    String scheme = getScheme(conf);
+    URI uri = new URI(scheme + "://hive@localhost:" + port + "/" + servletPath + "/" + NS);
     String jwt = generateJWT();
-    return new JSonClient(jwt, uri);
+    return new JsonClient(jwt, uri);
   }
 
   /**
    * A property client that uses http as transport.
    */
   @SuppressWarnings("unchecked")
-  public class JSonClient implements HttpPropertyClient {
+  public class JsonClient implements HttpPropertyClient {
     private final URI uri;
     private final String jwt;
-    JSonClient(String token, URI uri) {
+    JsonClient(String token, URI uri) {
       this.jwt = token;
       this.uri = uri;
     }
@@ -167,84 +151,7 @@ public class HMSServletTest extends HMSTestBase {
     }
   }
 
-  @Test
-  public void testServletEchoA() throws Exception {
-    String path = MetastoreConf.getVar(conf, MetastoreConf.ConfVars.PROPERTIES_SERVLET_PATH);
-    String scheme = MetastoreConf.getBoolVar(conf, MetastoreConf.ConfVars.HTTPSERVER_USE_HTTPS)
-            ? "https" : "http";
-    URI uri = new URI(scheme + "://hive@localhost:" + servletPort + "/" + path + "/" + NS);
-    Map<String, String> json = Collections.singletonMap("method", "echo");
-    String jwt = generateJWT();
-    // succeed
-    Object response = clientCall(jwt, uri, "POST", json);
-    Assert.assertNotNull(response);
-    Assert.assertEquals(json, response);
-    // fail (bad jwt)
-    String badJwt = generateJWT(jwtUnauthorizedKeyFile.toPath());
-    response = clientCall(badJwt, uri, "POST", json);
-    Assert.assertNull(response);
-  }
-
-  @Test
-  public void testProperties1() throws Exception {
-      runOtherProperties1(client);
-  }
-
-
-  @Test
-  public void testProperties0() throws Exception {
-    runOtherProperties0(client);
-  }
-
-  @Test
-  public void testProperties2() throws Exception {
-    HttpClient httpClient = createHttpClient(conf);
-    String jwt = generateJWT();
-    NameValuePair[] nvp = new NameValuePair[]{
-        new BasicNameValuePair("key", "db0.table01.fillFactor"),
-        new BasicNameValuePair("key", "db0.table04.fillFactor")
-    };
-    String scheme = MetastoreConf.getBoolVar(conf, MetastoreConf.ConfVars.HTTPSERVER_USE_HTTPS)
-            ? "https" : "http";
-    URI uri = new URIBuilder()
-        .setScheme(scheme)
-        .setUserInfo("hive")
-        .setHost("localhost")
-        .setPort(servletPort)
-        .setPath("/" + path + "/" + NS)
-        .setParameters(nvp)
-        .build();
-    HttpGet get = new HttpGet(uri);
-    get.addHeader("Authorization", "Bearer " + jwt);
-    get.addHeader("Content-Type", "application/json");
-    get.addHeader("Accept", "application/json");
-    get.addHeader(MetaStoreUtils.USER_NAME_HTTP_HEADER, "hive");
-
-    Map<String,String> result = null;
-    HttpResponse response = httpClient.execute(get);
-    try {
-      Assert.assertEquals(HttpServletResponse.SC_OK, response.getStatusLine().getStatusCode());
-      HttpEntity entity = response.getEntity();
-      if (entity != null) {
-        Gson gson = new GsonBuilder().create();
-        ContentType contentType = ContentType.getOrDefault(entity);
-        Charset charset = contentType.getCharset();
-        Reader reader = new InputStreamReader(entity.getContent(), charset);
-        result = (Map<String, String>) gson.fromJson(reader, Object.class);
-      }
-      Assert.assertNotNull(result);
-      Assert.assertEquals(2, result.size());
-    } finally {
-      if (response instanceof AutoCloseable) {
-        ((AutoCloseable) response).close();
-      }
-      if (httpClient instanceof AutoCloseable) {
-        ((AutoCloseable) httpClient).close();
-      }
-    }
-  }
-
- /**
+  /**
    * Reads the content of a Reader into a String.
    * @param reader the reader to read from
    * @return the content as a String
@@ -306,12 +213,37 @@ public class HMSServletTest extends HMSTestBase {
     int responseCode = con.getResponseCode();
     if (responseCode == HttpServletResponse.SC_OK) {
       try (Reader reader = new BufferedReader(
-          new InputStreamReader(con.getInputStream(), StandardCharsets.UTF_8))) {
+              new InputStreamReader(con.getInputStream(), StandardCharsets.UTF_8))) {
         return new Gson().fromJson(reader, Object.class);
       }
     }
     return null;
   }
 
+  @Test
+  public void testEcho() throws Exception {
+    String path = MetastoreConf.getVar(conf, MetastoreConf.ConfVars.PROPERTIES_SERVLET_PATH);
+    String scheme = getScheme(conf);
+    URI uri = new URI(scheme + "://hive@localhost:" + servletPort + "/" + path + "/" + NS);
+    Map<String, String> json = Collections.singletonMap("method", "echo");
+    String jwt = generateJWT();
+    // succeed
+    Object response = clientCall(jwt, uri, "POST", json);
+    Assert.assertNotNull(response);
+    Assert.assertEquals(json, response);
+    // fail (bad jwt)
+    String badJwt = generateJWT(jwtUnauthorizedKeyFile.toPath());
+    response = clientCall(badJwt, uri, "POST", json);
+    Assert.assertNull(response);
+  }
 
+  @Test
+  public void testProperties1() throws Exception {
+      runOtherProperties1(client);
+  }
+
+  @Test
+  public void testProperties0() throws Exception {
+    runOtherProperties0(client);
+  }
 }
